Introduces an explicit interface for Arcs debugging tools
