+++ src/devtools-connector/devtools-connection.ts
@@ -8,9 +8,9 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import {DevtoolsBroker} from '../../../devtools/shared/devtools-broker.js';
-import {assert} from '../../platform/assert-web.js';
-import {DevtoolsChannel} from '../../platform/devtools-channel-web.js';
+import {DevtoolsBroker} from '../../devtools/shared/devtools-broker.js';
+import {assert} from '../platform/assert-web.js';
+import {DevtoolsChannel} from '../platform/devtools-channel-web.js';
 import {DevtoolsChannelStub} from './testing/devtools-channel-stub.js';
 
 let channel: DevtoolsChannel = null;
