+++ src/devtools-connector/tracing-adapter.ts
@@ -7,7 +7,7 @@
  * subject to an additional IP rights grant found at
  * http://polymer.github.io/PATENTS.txt
  */
-import {Tracing} from '../../tracelib/trace.js';
+import {Tracing} from '../tracelib/trace.js';
 import {ArcDevtoolsChannel} from './abstract-devtools-channel.js';
 
 let streamingToDevtools = false;
