+++ src/planning/debug/planning-explorer-adapter.ts
@@ -7,10 +7,9 @@
  * subject to an additional IP rights grant found at
  * http://polymer.github.io/PATENTS.txt
  */
-import {DevtoolsConnection} from '../../runtime/debug/devtools-connection.js';
+import {DevtoolsConnection} from '../../devtools-connector/devtools-connection.js';
 import {Trigger} from '../plan/plan-producer.js';
 import {Planificator} from '../plan/planificator.js';
-import {PlanningResult} from '../plan/planning-result.js';
 import {Suggestion} from '../plan/suggestion';
 
  export class PlanningExplorerAdapter {
