+++ src/planning/recipe-index.ts
@@ -11,7 +11,7 @@
 import {Strategizer, Strategy} from '../planning/strategizer.js';
 import {assert} from '../platform/assert-web.js';
 import {Arc} from '../runtime/arc.js';
-import {DevtoolsConnection} from '../runtime/debug/devtools-connection.js';
+import {DevtoolsConnection} from '../devtools-connector/devtools-connection.js';
 import {Manifest} from '../runtime/manifest.js';
 import {Modality} from '../runtime/modality.js';
 import {ProvideSlotConnectionSpec, ConsumeSlotConnectionSpec} from '../runtime/particle-spec.js';
