+++ src/planning/speculator.ts
@@ -10,7 +10,7 @@
 
 import {assert} from '../platform/assert-web.js';
 import {Arc} from '../runtime/arc.js';
-import {DevtoolsConnection} from '../runtime/debug/devtools-connection.js';
+import {DevtoolsConnection} from '../devtools-connector/devtools-connection.js';
 import {Description} from '../runtime/description.js';
 import {Recipe} from '../runtime/recipe/recipe.js';
 import {Relevance} from '../runtime/relevance.js';
