+++ src/runtime/api-channel.ts
@@ -11,8 +11,7 @@
 import {assert} from '../platform/assert-web.js';
 
 import {Arc} from './arc.js';
-import {DevtoolsConnection} from './debug/devtools-connection.js';
-import {OuterPortAttachment} from './debug/outer-port-attachment.js';
+import {ArcInspector} from './arc-inspector.js';
 import {Handle} from './handle.js';
 import {ParticleSpec} from './particle-spec.js';
 import {Particle} from './particle.js';
@@ -192,15 +191,15 @@ class ThingMapper {
 export class APIPort {
   private _port: MessagePort;
   _mapper: ThingMapper;
-  protected _debugAttachment: OuterPortAttachment;
-  protected _attachStack: boolean;
+  protected inspector: ArcInspector;
+  protected attachStack: boolean;
   messageCount: number;
   constructor(messagePort: MessagePort, prefix: string) {
     this._port = messagePort;
     this._mapper = new ThingMapper(prefix);
     this._port.onmessage = async e => this._processMessage(e);
-    this._debugAttachment = null;
-    this._attachStack = false;
+    this.inspector = null;
+    this.attachStack = false;
     this.messageCount = 0;
 
     this._testingHook();
@@ -217,17 +216,17 @@ export class APIPort {
   async _processMessage(e) {
     assert(this['before' + e.data.messageType] !== undefined);
     const count = this.messageCount++;
-    if (this._debugAttachment) {
-      this._debugAttachment.handlePecMessage('on' + e.data.messageType, e.data.messageBody, count, e.data.stack);
+    if (this.inspector) {
+      this.inspector.pecMessage('on' + e.data.messageType, e.data.messageBody, count, e.data.stack);
     }
     this['before' + e.data.messageType](e.data.messageBody);
   }
 
   send(name: string, args: {}) {
-    const call = {messageType: name, messageBody: args, stack: this._attachStack ? new Error().stack : undefined};
+    const call = {messageType: name, messageBody: args, stack: this.attachStack ? new Error().stack : undefined};
     const count = this.messageCount++;
-    if (this._debugAttachment) {
-      this._debugAttachment.handlePecMessage(name, args, count, new Error().stack);
+    if (this.inspector) {
+      this.inspector.pecMessage(name, args, count, new Error().stack);
     }
     this._port.postMessage(call);
   }
@@ -401,10 +400,10 @@ function AutoConstruct<S extends {prototype: {}}>(target: S) {
 export abstract class PECOuterPort extends APIPort {
   constructor(messagePort: MessagePort, arc: Arc) {
     super(messagePort, 'o');
-    DevtoolsConnection.onceConnected.then(devtoolsChannel => {
-      this.DevToolsConnected();
-      this._debugAttachment = new OuterPortAttachment(arc, devtoolsChannel);
-    });
+    this.inspector = arc.inspector;
+    if (this.inspector) {
+      this.inspector.onceActive.then(() => this.DevToolsConnected());
+    }
   }
 
   @NoArgs Stop() {}
@@ -522,6 +521,6 @@ export abstract class PECInnerPort extends APIPort {
   // send it along with the message. We only want to do this after a DevTools connection has been detected, which
   // we can't directly detect inside a worker context, so the PECOuterPort will send an API message instead.
   onDevToolsConnected() {
-    this._attachStack = true;
+    this.attachStack = true;
   }
 }
