+++ src/runtime/arc.ts
@@ -10,8 +10,7 @@
 
 import {assert} from '../platform/assert-web.js';
 
-import {ArcDebugListenerDerived} from './debug/abstract-devtools-channel.js';
-import {ArcDebugHandler} from './debug/arc-debug-handler.js';
+import {ArcInspector, ArcInspectorFactory} from './arc-inspector.js';
 import {FakePecFactory} from './fake-pec-factory.js';
 import {Id, IdGenerator, ArcId} from './id.js';
 import {Loader} from './loader.js';
@@ -44,7 +43,7 @@ type ArcOptions = Readonly<{
   speculative?: boolean;
   innerArc?: boolean;
   stub?: boolean
-  listenerClasses?: ArcDebugListenerDerived[];
+  inspectorFactory?: ArcInspectorFactory
 }>;
 
 type DeserializeArcOptions = Readonly<{
@@ -54,7 +53,7 @@ type DeserializeArcOptions = Readonly<{
   loader: Loader;
   fileName: string;
   context: Manifest;
-  listenerClasses?: ArcDebugListenerDerived[];
+  inspectorFactory?: ArcInspectorFactory
 }>;
 
 type SerializeContext = {handles: string, resources: string, interfaces: string, dataResources: Map<string, string>};
@@ -81,9 +80,9 @@ export class Arc {
   // Map from each store to its description (originating in the manifest).
   private readonly storeDescriptions = new Map<StorageProviderBase, string>();
   private waitForIdlePromise: Promise<void> | null;
-  private readonly debugHandler: ArcDebugHandler;
+  private readonly inspectorFactory: ArcInspectorFactory;
+  public readonly inspector: ArcInspector;
   private readonly innerArcsByParticle: Map<Particle, Arc[]> = new Map();
-  private readonly listenerClasses: ArcDebugListenerDerived[];
   private readonly instantiateMutex = new Mutex();
 
   readonly id: Id;
@@ -91,7 +90,7 @@ export class Arc {
   loadedParticleInfo = new Map<string, {spec: ParticleSpec, stores: Map<string, StorageProviderBase>}>();
   readonly pec: ParticleExecutionHost;
 
-  constructor({id, context, pecFactory, slotComposer, loader, storageKey, storageProviderFactory, speculative, innerArc, stub, listenerClasses} : ArcOptions) {
+  constructor({id, context, pecFactory, slotComposer, loader, storageKey, storageProviderFactory, speculative, innerArc, stub, inspectorFactory} : ArcOptions) {
     // TODO: context should not be optional.
     this._context = context || new Manifest({id});
     // TODO: pecFactory should not be optional. update all callers and fix here.
@@ -110,15 +109,13 @@ export class Arc {
     this.isInnerArc = !!innerArc; // undefined => false
     this.isStub = !!stub;
     this._loader = loader;
-
+    this.inspectorFactory = inspectorFactory;
+    this.inspector = inspectorFactory && inspectorFactory.create(this);
     this.storageKey = storageKey;
-
     const pecId = this.generateID();
     const innerPecPort = this.pecFactory(pecId, this.idGenerator);
     this.pec = new ParticleExecutionHost(innerPecPort, slotComposer, this);
     this.storageProviderFactory = storageProviderFactory || new StorageProviderFactory(this.id);
-    this.listenerClasses = listenerClasses;
-    this.debugHandler = new ArcDebugHandler(this, listenerClasses);
   }
 
   get loader(): Loader {
@@ -197,7 +194,7 @@ export class Arc {
 
   createInnerArc(transformationParticle: Particle): Arc {
     const id = this.generateID('inner');
-    const innerArc = new Arc({id, pecFactory: this.pecFactory, slotComposer: this.pec.slotComposer, loader: this._loader, context: this.context, innerArc: true, speculative: this.isSpeculative, listenerClasses: this.listenerClasses});
+    const innerArc = new Arc({id, pecFactory: this.pecFactory, slotComposer: this.pec.slotComposer, loader: this._loader, context: this.context, innerArc: true, speculative: this.isSpeculative, inspectorFactory: this.inspectorFactory});
 
     let particleInnerArcs = this.innerArcsByParticle.get(transformationParticle);
     if (!particleInnerArcs) {
@@ -377,7 +374,7 @@ ${this.activeRecipe.toString()}`;
     await store.set(arcInfoType.newInstance(this.id, serialization));
   }
 
-  static async deserialize({serialization, pecFactory, slotComposer, loader, fileName, context, listenerClasses}: DeserializeArcOptions): Promise<Arc> {
+  static async deserialize({serialization, pecFactory, slotComposer, loader, fileName, context, inspectorFactory}: DeserializeArcOptions): Promise<Arc> {
     const manifest = await Manifest.parse(serialization, {loader, fileName, context});
     const arc = new Arc({
       id: Id.fromString(manifest.meta.name),
@@ -387,7 +384,7 @@ ${this.activeRecipe.toString()}`;
       loader,
       storageProviderFactory: manifest.storageProviderFactory,
       context,
-      listenerClasses
+      inspectorFactory
     });
     await Promise.all(manifest.stores.map(async store => {
       const tags = manifest.storeTags.get(store);
@@ -458,7 +455,7 @@ ${this.activeRecipe.toString()}`;
                          loader: this._loader,
                          speculative: true,
                          innerArc: this.isInnerArc,
-                         listenerClasses: this.listenerClasses});
+                         inspectorFactory: this.inspectorFactory});
     const storeMap: Map<StorageProviderBase, StorageProviderBase> = new Map();
     for (const store of this._stores) {
       const clone = await arc.storageProviderFactory.construct(store.id, store.type, 'volatile');
@@ -604,8 +601,10 @@ ${this.activeRecipe.toString()}`;
       // TODO: pass slot-connections instead
       await this.pec.slotComposer.initializeRecipe(this, particles);
     }
-    
-    this.debugHandler.recipeInstantiated({particles, activeRecipe: this.activeRecipe.toString()});
+
+    if (this.inspector) {
+      this.inspector.recipeInstantiated(particles, this.activeRecipe.toString());
+    }
   }
 
   async createStore(type: Type, name?: string, id?: string, tags?: string[], storageKey: string = undefined): Promise<StorageProviderBase> {
